requirements.txt
Django>=2.2,<2.3
gunicorn==20.0.4
dj-database-url
dj-static
djangorestframework
psycopg2
python-decouple